+++ src/gui/settingsDialog/ImageGrabberSettings.h
@@ -28,7 +28,6 @@
 #include "backend/config/KsnipConfig.h"
 #include "src/widgets/ColorButton.h"
 #include "src/widgets/NumericComboBox.h"
-#include "src/common/provider/ScaledSizeProvider.h"
 
 class ImageGrabberSettings : public QGroupBox
 {