+++ src/gui/settingsDialog/ScriptUploaderSettings.h
@@ -29,7 +29,6 @@
 #include <QFileDialog>
 
 #include "src/backend/config/KsnipConfig.h"
-#include "src/common/provider/ScaledSizeProvider.h"
 
 class ScriptUploaderSettings : public QGroupBox
 {