+++ src/widgets/MainToolBar.cpp
@@ -65,13 +65,13 @@ MainToolBar::MainToolBar(const QList<CaptureModes> &captureModes, QAction* undoA
     mCropButton->addAction(mCropAction);
     mCropButton->setDefaultAction(mCropAction);
 
-	auto clockPixmap = IconLoader::load(QStringLiteral("clock.svg")).pixmap(ScaledSizeProvider::getScaledSize(QSize(24, 24)));
+	auto clockPixmap = IconLoader::load(QStringLiteral("clock.svg")).pixmap(QSize(24, 24));
 	mDelayLabel->setPixmap(clockPixmap);
 	mDelayLabel->setContentsMargins(0, 0, 2, 0);
 	mDelayLabel->setToolTip(tr("Delay in seconds between triggering\n"
 	                           "and capturing screenshot."));
     mDelayPicker->setSuffix(tr("s"));
-	mDelayPicker->setFixedWidth(ScaledSizeProvider::getScaledWidth(55));
+	mDelayPicker->setFixedWidth(55);
 	mDelayPicker->setToolTip(mDelayLabel->toolTip());
     connect(mDelayPicker, &CustomSpinBox::valueChanged, this, &MainToolBar::captureDelayChanged);